refactor(domain): split persistence module into submodules
