+++ crates/luban_domain/src/persistence/mod.rs
@@ -1,0 +1,6 @@
+mod fonts;
+mod load;
+mod save;
+
+pub(crate) use load::apply_persisted_app_state;
+pub(crate) use save::to_persisted_app_state;
